refactor: move PromptAccountModal to a separate file
