+++ src/components/SubmitModal/SubmitModal.tsx
@@ -1,20 +1,9 @@
 "use client";
 
-import {
-  currencyIsStables$,
-  estimatedCost$,
-  priceTotals$,
-  signerBalance$,
-} from "@/components/RfpForm/data";
-import { TOKEN_DECIMALS } from "@/constants";
-import { SelectGroup } from "@radix-ui/react-select";
+import { currencyIsStables$ } from "@/components/RfpForm/data";
 import { state, useStateObservable } from "@react-rxjs/core";
-import { createSignal } from "@react-rxjs/utils";
-import { CheckCircle2, TriangleAlert } from "lucide-react";
-import { AccountId } from "polkadot-api";
 import { useEffect } from "react";
 import {
-  combineLatest,
   filter,
   map,
   merge,
@@ -24,35 +13,20 @@ import {
   take,
   withLatestFrom,
 } from "rxjs";
-import { formValue$ } from "../RfpForm/data/formValue";
-import { bounties$, bountyById$ } from "../RfpForm/FundingBountyCheck";
 import { selectedAccount$ } from "../SelectAccount";
-import { PickExtension } from "../SelectAccount/PickExtension";
-import { PickExtensionAccount } from "../SelectAccount/PickExtensionAccount";
 import { Loading } from "../Spinner";
-import { Button } from "../ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "../ui/dialog";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectLabel,
-  SelectTrigger,
-  SelectValue,
-} from "../ui/select";
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../ui/dialog";
 import {
   dismiss,
   dismiss$,
   formDataChange$,
-  submit,
   submittedFormData$,
 } from "./modalActions";
+import {
+  accountSelected$,
+  PromptAccountModal,
+  signerStepValidity$,
+} from "./PromptAccountModal";
 import { SubmitBountyModal } from "./SubmitBountyModal";
 import { SubmitMultisigRfpModal } from "./SubmitMultisigRfpModal";
 
@@ -65,70 +39,6 @@ type ModalView = null | {
     | "multisig_transaction_steps";
 };
 
-const signerStepValidity$ = state(
-  combineLatest({
-    selectedAccount: selectedAccount$,
-    signerBalance: signerBalance$,
-    estimatedCost: estimatedCost$,
-    priceTotals: priceTotals$,
-    formValue: formValue$,
-    bounty: formValue$.pipe(
-      switchMap((form) =>
-        form.isChildRfp && form.parentBountyId != null
-          ? bountyById$(form.parentBountyId)
-          : [null],
-      ),
-    ),
-  }).pipe(
-    map(
-      ({
-        selectedAccount,
-        signerBalance,
-        estimatedCost,
-        priceTotals,
-        formValue,
-        bounty,
-      }) => {
-        const bountyBalance = bounty?.balance ?? null;
-        const bountySigners = bounty?.signers ?? null;
-        const accountGenericSs58 = selectedAccount
-          ? AccountId().dec(AccountId().enc(selectedAccount.address))
-          : null;
-
-        const feesAndDeposits =
-          signerBalance && estimatedCost
-            ? estimatedCost.deposits + estimatedCost.fees < signerBalance
-            : false;
-        const totalAmount = priceTotals
-          ? BigInt(priceTotals.totalAmountWithBuffer * 10 ** TOKEN_DECIMALS)
-          : null;
-        const parentBounty = formValue.isChildRfp
-          ? !!bountyBalance &&
-            !!bountySigners &&
-            !!accountGenericSs58 &&
-            !!totalAmount &&
-            bountyBalance > totalAmount &&
-            bountySigners.includes(accountGenericSs58)
-          : true;
-
-        return {
-          isValid: feesAndDeposits && parentBounty,
-          signerBalance,
-          estimatedCost,
-          priceTotals,
-          totalAmount,
-          accountGenericSs58,
-          bountyBalance,
-          bountySigners,
-        };
-      },
-    ),
-  ),
-  null,
-);
-
-const [accountSelected$, onAccountSelected] = createSignal();
-// This observable manages the modal's view state before transaction steps
 const submitModal$ = state(
   merge(formDataChange$, dismiss$.pipe(map(() => null))).pipe(
     withLatestFrom(selectedAccount$),
@@ -170,158 +80,6 @@ const submitModal$ = state(
   ),
   null,
 );
-
-const PromptAccountModal = () => {
-  const formValue = useStateObservable(formValue$);
-  const validity = useStateObservable(signerStepValidity$);
-
-  const renderSignerBalanceCheck = () => {
-    if (!validity?.accountGenericSs58 || !validity?.estimatedCost) return null;
-
-    if (validity.signerBalance === null) {
-      return (
-        <div className="py-8 flex flex-col justify-center items-center space-y-2">
-          <Loading />
-          <p className="text-sm text-muted-foreground">
-            Checking your available balance...
-          </p>
-        </div>
-      );
-    }
-
-    if (
-      validity.signerBalance >
-      validity.estimatedCost.deposits + validity.estimatedCost.fees
-    ) {
-      return (
-        <div className="poster-alert alert-success flex items-center gap-3">
-          <CheckCircle2 size={20} className="shrink-0 text-lilypad" />
-          <div className="text-sm">
-            <strong>Nice:</strong> you have enough balance to launch the RFP 🚀
-          </div>
-        </div>
-      );
-    }
-
-    return (
-      <div className="poster-alert alert-error flex items-center gap-3">
-        <TriangleAlert size={20} className="shrink-0" />
-        <div className="text-sm">
-          <strong>Uh-oh:</strong> not enough balance. Please add funds or select
-          another wallet.
-        </div>
-      </div>
-    );
-  };
-
-  return (
-    <Dialog open onOpenChange={(isOpen) => (isOpen ? null : dismiss())}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Connect Wallet</DialogTitle>
-          <DialogDescription>
-            Please connect a wallet to submit the RFP.
-          </DialogDescription>
-        </DialogHeader>
-        <PickExtension />
-        <PickExtensionAccount autoSelect />
-        {renderSignerBalanceCheck()}
-        {formValue.isChildRfp ? <PromptParentBounty /> : null}
-        <div className="text-right">
-          <Button disabled={!validity?.isValid} onClick={onAccountSelected}>
-            Next
-          </Button>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const PromptParentBounty = () => {
-  const formValue = useStateObservable(submittedFormData$);
-  const bounties = useStateObservable(bounties$);
-  const validity = useStateObservable(signerStepValidity$);
-
-  if (!formValue) return null;
-
-  const signerBounties = bounties.filter((b) =>
-    // eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-    b.signers.includes(validity?.accountGenericSs58!),
-  );
-
-  const value = signerBounties.some((v) => v.id === formValue.parentBountyId)
-    ? formValue.parentBountyId
-    : null;
-
-  const renderBountyBalanceCheck = () => {
-    if (!validity?.bountySigners || !validity.totalAmount) return null;
-
-    if (validity.bountyBalance === null) {
-      return (
-        <div className="py-8 flex flex-col justify-center items-center space-y-2">
-          <Loading />
-          <p className="text-sm text-muted-foreground">
-            Checking for the bounty balance balance...
-          </p>
-        </div>
-      );
-    }
-
-    if (validity.bountyBalance > validity.totalAmount) {
-      return (
-        <div className="poster-alert alert-success flex items-center gap-3">
-          <CheckCircle2 size={20} className="shrink-0 text-lilypad" />
-          <div className="text-sm">
-            The bounty has enough balance for the child bounty!
-          </div>
-        </div>
-      );
-    }
-
-    return (
-      <div className="poster-alert alert-error flex items-center gap-3">
-        <TriangleAlert size={20} className="shrink-0" />
-        <div className="text-sm">
-          <strong>Uh-oh:</strong> the selected bounty doesn't have enough
-          balance.
-        </div>
-      </div>
-    );
-  };
-
-  return (
-    <div className="space-y-2">
-      <p className="text-sm">Select the parent bounty</p>
-      <Select
-        value={String(value ?? "")}
-        onValueChange={(v) =>
-          submit({ ...formValue, parentBountyId: Number(v) })
-        }
-      >
-        <SelectTrigger className="w-full data-[size=default]:h-auto">
-          <SelectValue placeholder="Choose a bounty" />
-        </SelectTrigger>
-        <SelectContent>
-          <SelectGroup>
-            {signerBounties.length ? (
-              signerBounties.map((bounty) => (
-                <SelectItem key={bounty.id} value={String(bounty.id)}>
-                  {bounty.id}. {bounty.description}
-                </SelectItem>
-              ))
-            ) : (
-              <SelectLabel className="font-bold">
-                The account you selected doesn't seem to have a curator role for
-                any of the bounties.
-              </SelectLabel>
-            )}
-          </SelectGroup>
-        </SelectContent>
-      </Select>
-      {renderBountyBalanceCheck()}
-    </div>
-  );
-};
 
 export const SubmitModal = () => {
   const modalStatus = useStateObservable(submitModal$);
